fix(YTMusic): resolve 12 build errors in YouTubeMusic provider
- Change YTMusicClient.auth from private to internal for extension access
- Make YTMusicAuth, AuthState, BridgeCodeResponse public for provider interface
- Make YTMusicError public with protocol-required properties
- Replace non-existent BaseItemDto.externalUrls with providerIDs
- Fix Int64 -> Int type mismatch in runTimeTicks assignment
- Change DeviceCodeResponse to BridgeCodeResponse in provider return type
- Cast Int to CGFloat for imageSource maxWidth parameter

+++ Shared/Services/DownloadTask.swift
@@ -163,11 +163,11 @@ class DownloadTask: NSObject, ObservableObject {
 
         switch type {
         case .movie, .series:
-            guard let url = item.imageSource(imageType, maxWidth: maxWidth).url else { return }
+            guard let url = item.imageSource(imageType, maxWidth: CGFloat(maxWidth)).url else { return }
             imageURL = url
         case .episode where imageType == .backdrop:
             // Episodes use primary image for backdrop display
-            guard let url = item.imageSource(.primary, maxWidth: maxWidth).url else { return }
+            guard let url = item.imageSource(.primary, maxWidth: CGFloat(maxWidth)).url else { return }
             imageURL = url
         default:
             return
